added EnemyCar.java and EnemyInput.java. this will not be implemented at the moment

+++ src/com/akivaGrobman/cars/Car.java
@@ -6,9 +6,9 @@ import java.awt.image.BufferedImage;
 
 public abstract class Car {
 
-    protected float x;
-    protected float y;
-    protected float angle;
+    protected double x;
+    protected double y;
+    protected double angle;
     protected BufferedImage carImage;
 
     public abstract void tick();
